Refactors to a Checker class
