+++ lib/cc/cli.rb
@@ -6,11 +6,13 @@ require "cc/workspace"
 require "cc/yaml"
 
 require "cc/cli/analyze"
+require "cc/cli/checker"
 require "cc/cli/command"
 require "cc/cli/console"
 require "cc/cli/engines"
 require "cc/cli/help"
 require "cc/cli/init"
+require "cc/cli/output"
 require "cc/cli/prepare"
 require "cc/cli/runner"
 require "cc/cli/test"
